requirements.txt
ttkbootstrap>=1.10.1
reportlab>=4.1.0
matplotlib>=3.8.0
Pillow>=10.3.0